requirements.txt
python-telegram-bot==20.8
aiosqlite==0.19.0
pytesseract==0.3.10
Pillow==10.3.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0


